initial division of plot_figure code re #186
